# lib/api-zod/src/generated/types/createCreativeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SparqForge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCreativeInputSelectedAssetsItem } from "./createCreativeInputSelectedAssetsItem";

export interface CreateCreativeInput {
  brandId: string;
  templateId?: string | null;
  name: string;
  briefText?: string | null;
  referenceUrl?: string | null;
  selectedAssets: CreateCreativeInputSelectedAssetsItem[];
  selectedHashtagSets?: string[] | null;
  createdBy: string;
}